cleanup dry run mode

+++ 04_cleanup.py
@@ -1,6 +1,12 @@
 import h5py
 import glob
 import os
+from sys import argv
+
+if len(argv)>1:
+    dry_run = True
+else:
+    dry_run = False
 
 files_traj1d = glob.glob("data/traj1d/*.h5py")
 
@@ -10,20 +16,25 @@ for infile in files_traj1d:
 
         N = myfile.attrs["N"]
 
+        myfile.close()
+
         print("Ok... ", infile, N)
 
     except RuntimeError as e:
         print("Removing... ", infile)
         print(e)
-        os.remove(infile)
+        if not dry_run:
+            os.remove(infile)
 
     except OSError as e:
         print("Removing... ",infile)
         print(e)
-        os.remove(infile)
+        if not dry_run:
+            os.remove(infile)
 
     except KeyError as e:
         print("Removing... ",infile)
         print(e)
-        os.remove(infile)
+        if not dry_run:
+            os.remove(infile)
